% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.mini.chains.R
\name{combine.mini.chains}
\alias{combine.mini.chains}
\title{combine.mini.chains}
\usage{
combine.mini.chains(mini.chains)
}
\arguments{
\item{mini.chains}{a \code{"mini.chains"} objects or the path to a mini.chains object}
}
\description{
Combine the mini chains results into a big MCMCglmm object (MCMCglmm)
}
\author{
Thomas Guillerme
}
