% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.VCV.R
\name{convert.VCV}
\alias{convert.VCV}
\title{Convert VCV}
\usage{
convert.VCV(input, level = 0.95)
}
\arguments{
\item{input}{either a VCV list, a VCV matrix or a matrix of major axis coordinates.}

\item{level}{which confidence interval level to use}
}
\value{
If the input is a matrix, returns a VCV list.
If the input is a VCV list or a VCV matrix, returns a major axis coordinates matrix.
}
\description{
convert a VCV matrix into coordinates
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
