% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcmcglmmm-package.R
\name{demo_data}
\alias{demo_data}
\alias{demo_tree}
\title{Demo}
\format{
one \code{data.frame}, one \code{phylo}.
}
\description{
A small demo dataset (a \code{data.frame} and a \code{phylo} object).
}
\details{
This dataset is based on a random subset of 55 Charadriiformes (gulls, plovers and sandpipers) from Cooney et al 2017.
It contains:
\itemize{
\item \code{demo_data} A \code{"data.frame"} .
\item \code{demo_tree} A consensus tree of 55 charadriiformes species (\code{"phylo"}).
}
}
\references{
Cooney CR, Bright JA, Capp EJ, Chira AM, Hughes EC, Moody CJ, Nouri LO, Varley ZK, Thomas GH. Mega-evolutionary dynamics of the adaptive radiation of birds. Nature. 2017 Feb;542(7641):344-7.
}
