% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.mini.chains.R
\name{diagnose.mini.chains}
\alias{diagnose.mini.chains}
\title{diagnose.mini.chains}
\usage{
diagnose.mini.chains(mini.chains, what = "eff.samp", plot = TRUE, ...)
}
\arguments{
\item{mini.chains}{a mini chain MCMCglmm result}

\item{what}{if mini.chains is a single chain, which part from \code{\link[MCMCglmm]{summary.MCMCglmm}} to diagnose (default is \code{"eff.samp"} for the effective sample size).}

\item{plot}{if mini.chains is a single chain, logical, whether to plot the results (\code{TRUE}; default) or not (\code{FALSE}).}

\item{...}{if mini.chains is a single chain: if \code{plot = TRUE}, any optional arguments to be past to \code{\link[graphics]{hist}}. Else, any optional arguments to be passed to \code{\link[coda]{gelman.diag}}.}
}
\description{
Diagnose a set of mini chains results
}
\details{
If the input is a list of chains straight from \code{run.mini.chains}, a Gelman and Rubin convergence diagnosis is run (\code{\link[coda]{gelman.diag}}). Else if the input is a single chain combined from \code{combine.mini.chains} only the effective sample size for each parameter (or \code{what}) is measured.
}
\author{
Thomas Guillerme
}
