% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcmcglmmm-package.R
\docType{package}
\name{dispRity-package}
\alias{dispRity-package}
\title{Running the mini chains MCMCglmm method}
\description{
A wrapper package with functions for running the mini chains MCMCglmm method.
}
\author{
Thomas Guillerme \href{mailto:guillert@tcd.ie}{guillert@tcd.ie}, Andrew Beckerman, Natalie Cooper, Gavin Thomas.
}
\concept{MCMCglmm}
