% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.parameters.R
\name{extract.parameters}
\alias{extract.parameters}
\title{Extract parameters}
\usage{
extract.parameters(
  chains,
  parameters = c("burnin", "priors"),
  buffer = 0.1,
  nu = 0.05
)
}
\arguments{
\item{chains}{The prior MCMCglmm chains}

\item{parameters}{The parameters to extract (see details).}

\item{buffer}{The buffer for the burnin (how many itteration to include past the first median estimate)}

\item{nu}{The degree of belief parameter (nu) for the priors.}
}
\description{
Extracting parameters from prior MCMCglmm chains
}
\details{
The parameters that can be extracted are:
\itemize{
\item \code{"burnin"}: the global burnin length
\item \code{"priors"}: a list of priors (from the posteriors)
}
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
