% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.prior.R
\name{flat.prior}
\alias{flat.prior}
\title{Making flat priors}
\usage{
flat.prior(residuals = 1, randoms = 1, ntraits, nu = 0)
}
\arguments{
\item{residuals}{the number of residual terms (R structure in the list)}

\item{randoms}{the number of random terms (G structure in the list)}

\item{ntraits}{the number of traits}

\item{nu}{the value for nu (applied throughout)}
}
\description{
Making a list of flat priors to be fed to a mini.chain
}
\examples{
## Default flat priors for 2 traits
flat.prior(ntraits = 2)

## Flat priors for a model with 1 residual and 4 random terms
## for 3 traits and a nu of 0.002
flat.prior(residuals = 1, randoms = 4, ntraits = 3, nu = 0.002)

}
\author{
Thomas Guillerme
}
