% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lambda.R
\name{get.lambda}
\alias{get.lambda}
\title{Get lambda}
\usage{
get.lambda(shape)
}
\arguments{
\item{shape}{the shape value}
}
\description{
Get the lambda value corresponding to the input shape
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
