% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.one.axis.R
\name{get.one.axis}
\alias{get.one.axis}
\title{Get the coordinates of one axis}
\usage{
get.one.axis(data, axis = 1, level = 0.95, dimensions)
}
\arguments{
\item{data}{a VCV matrix}

\item{axis}{which axis to calculate (1 is the major axis, 2 is the second major axis, etc...)}

\item{level}{the confidence interval level of the ellipse (default is 0.95).}

\item{dimensions}{optional, which dimensions to use.}
}
\description{
Get the coordinates of one axis from a VCV matrix
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
