% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.mini.chains.R
\name{make.mini.chains}
\alias{make.mini.chains}
\title{make.mini.chains}
\usage{
make.mini.chains(
  data,
  dimensions,
  tree,
  trait.family = "gaussian",
  residuals = "global",
  randoms = NULL,
  parameters,
  priors = 0.02,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{a full ordinated matrix with a column called \code{"animal"} and potentially one called \code{"clade"} (if needed).}

\item{dimensions}{a numeric or character vector of the dimensions to include.}

\item{tree}{the \code{"phylo"} or \code{"multiPhylo"} object to run the models.}

\item{trait.family}{the family of the traits (default is \code{"gaussian"}).}

\item{residuals}{optional, the type of residuals (see details).}

\item{randoms}{optional, the type of randoms (see details).}

\item{parameters}{the named list of parameters for the MCMCglmm (if missing, it is set to \code{list(nitt=1000, burnin = 100, thin = 100)} for a fast test).}

\item{priors}{A list of priors or a value of the overal nu parameter for generating flat priors for the \code{\link[MCMCglmm]{MCMCglmm}}.}

\item{verbose}{whether to make the \code{\link[MCMCglmm]{MCMCglmm}} verbose (\code{TRUE}; default) or not (\code{FALSE}).}

\item{...}{any other parameters to be passed to \code{\link[MCMCglmm]{MCMCglmm}}.}
}
\description{
Making a list of mini.chains object (MCMCglmm primers)
}
\details{
By default, no model is used for the random terms (\code{NULL}). The types of model for the residuals and random terms can be a combination of any of the following:
\itemize{
\item \code{"global"} for just an overall term effect. For the random terms that would be just a phylogenetic effect (\code{~us(trait):animal}) or for the residuals just an overall category (\code{~us(trait):unit}).
\item the name of the column(s) containing the factors for just as many levels of effects as there are non-empty levels in the data (\code{""} or \code{<NA>} are considered as empty levels). For example for the column \code{"clade"} the added random effects will be something like \code{us(at.level(clade,1):trait):animal} (i.e. a special phylogenetic random effect for clade 1) or \code{us(at.level(clade,1):trait):unit} for the residuals (i.e. a special residual effect for clade 1).
}
}
\author{
Thomas Guillerme
}
