% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.VCV.R
\name{make.VCV}
\alias{make.VCV}
\title{Make an ellipse}
\usage{
make.VCV(
  shape = 1,
  covariance = 0,
  rotation = 0,
  size = 1,
  position = 0,
  dimensions = 2,
  min.thick = 0
)
}
\arguments{
\item{shape}{the degree of roundness (1 = circle, 0 = line; default is 0.5)}

\item{covariance}{the covariance between dimensions (default is 0). Can be either a unique numeric value between 0 and 1 or a vector equal to the number of dimensions (see details).}

\item{rotation}{the rotation of the ellipse on each dimensions (expressed in degrees) (see details).}

\item{size}{the base radius of the circle at the base of the ellipse (default is 1).}

\item{position}{the ellipse centre's position (default is 0)}

\item{dimensions}{the number of dimensions (default is 2)}

\item{min.thick}{a proportional value to use a minimal thickness for the ellipse in each dimensions (default is 0). See details.}
}
\value{
A list of three elements:
$VCV: the variance covariance matrix of this ellipse
$loc: the location of the centre of the ellipse
$rotation: the rotation matrix (is null if rotation = 0)
}
\description{
Generates an var-covar matrix that can be plotted as an ellipse
}
\details{
Difference between \code{covariance} and \code{rotation}:
\itemize{
\item the \code{covariance} argument adds one or multiple covariance values to the VCV matrix (i.e. its upper and lower triangles). It does not change the orientation of the matrix per se but results in the matrice's ellipse to rotate up to 45 degree and change its thickness: for example a covariance of 1 will result into a 1D ellipse (straight line) at 45 degrees.
\item the \code{rotation} argument does not modify the values in the VCV matrix but outputs an additional $rotation argument that can be handled by plot.VCV. This changes the angle of the rotation matrix without changing its properties (size, shape, etc...).
}

When generating ellipses in low dimensions (e.g. < 10) or/and with extreme shape values, certain dimensions will be flat (i.e. will have a variance of 0). You can force dimensions to have a minimum thickness using \code{min.thick}. For example \code{min.thick = 0.1} will ensure that every dimensions has at least 10 percent of the required \code{size}.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
