% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.roundness.R
\name{plot.roundness}
\alias{plot.roundness}
\title{Plots a the roundness of a VCV matrix}
\usage{
\method{plot}{roundness}(VCV, dimensions, ...)
}
\arguments{
\item{VCV}{either a VCV list, a VCV matrix or a matrix of major axis coordinates.}

\item{dimensions}{which dimensions to use (if left empty, all are used).}

\item{...}{Any additional arguments to be passed to plot, lines or polygon.}
}
\description{
Displays the integral of the ordering of each major axes of an ellipse (i.e. it's roundness)
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
