% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VCV.R
\name{plot.VCV}
\alias{plot.VCV}
\title{Plots an ellipse}
\usage{
\method{plot}{VCV}(
  VCV,
  ellipse = TRUE,
  major.axes = TRUE,
  level = 0.95,
  dimensions = c(1, 2),
  VCV.vectors = FALSE,
  eigen.vectors = FALSE,
  eigen.values = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{VCV}{either a VCV list, a VCV matrix or a matrix of major axis coordinates.}

\item{ellipse}{logical, whether to plot the confidence interval ellipse or not (default is TRUE).}

\item{major.axes}{logical, whether to plot the major axes or not (default is TRUE).}

\item{level}{the confidence interval level to estimate the ellipse and the major axes (default is 0.95)}

\item{dimensions}{which dimensions to plot (up to 3!). Default is c(1,2).}

\item{VCV.vectors}{display the VCV matrix vectors (default is FALSE)}

\item{eigen.vectors}{display the eigen vectors (default is TRUE)}

\item{eigen.values}{if eigen.vectors = TRUE, whether to scale the eigen vectors by the eigen values.}

\item{add}{logical, whether to add it to an existing plot (TRUE) or not (FALSE; default)}

\item{...}{Any additional arguments to be passed to plot, lines or arrows.}
}
\description{
Plotting an ellipse from the output of make.ellipse
}
\details{
... can take specific options for specific elements to plot if you precede the plot argument by the element name. For example \code{col.eigen.vectors = "blue"} will make only the eigen.vectors blue.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
